[Streams] Small type changes for advanced mapping parameters (#211817)
## Summary

Switches to the `recursiveRecord` schema so we don't get these in
console:

```
[2025-02-19T15:47:07.556+01:00][WARN ][plugins.streams] Warning for PUT /api/streams/{name}: schema ZodUnknown at body.stream.ingest.wired.fields is not inspectable and could lead to runtime exceptions, convert it to a supported schema
[2025-02-19T15:47:07.557+01:00][WARN ][plugins.streams] Warning for POST /api/streams/{name}/schema/fields_simulation: schema ZodUnknown at body.field_definitions is not inspectable and could lead to runtime exceptions, convert it to a supported schema
[2025-02-19T15:47:07.557+01:00][WARN ][plugins.streams] Warning for POST /api/streams/{name}/processing/_simulate: schema ZodUnknown at body.detected_fields is not inspectable and could lead to runtime exceptions, convert it to a supported schema
```

I had to move the schema definition / types into another file otherwise
a circular dependency was introduced with the `fields/index.ts` file,
causing a `Cannot read properties of undefined (reading '_parse')`
error.

As far as I can see the `recursiveRecord` schema should handle / cover
the ES `MappingProperty` type fine.

(cherry picked from commit 670da35d4101fc64139876fbdf2aee5975036399)
