+++ x-pack/platform/packages/shared/kbn-streams-schema/src/models/index.ts
@@ -11,3 +11,4 @@ export * from './api';
 export * from './core';
 export * from './helpers';
 export * from './group';
+export * from './record_types';
